[#50] refactor : 백엔드 타입 정의
- 백엔드 타입 정의
- (이슈) controller.ts에서 Schema.create 할 때 타입 에러.
- 위 이슈 해결을 위한 임시 방편으로 model 객체의 타입을 any로 지정
